Initial commit
I am not the author of the code in this commit. This repository was created from
the public ChampSim code at
http://dropzone.tamu.edu/~cienlux/public/ChampSim_CRC2.tar.bz2 to track my
personal changes because I do not have access to the original git repository
used at Texas A&M for ChampSim development (which is located at
https://github.tamu.edu/cienlux/ChampSim.git, but requires a TAMU NetID to
access).

The changes in this repository may be rebased on top of the existing git history
later if it ever becomes available to me.